[UI] textEditor clean up
