+++ frontend/src/view/components/common/textEditor.tsx
@@ -1,28 +1,27 @@
 import * as React from 'react';
 import 'react-quill/dist/quill.snow.css';
-// import * as ReactQuill from 'react-quill';
 import ReactQuill from 'react-quill';
-import { bbcode2html, html2bbcode, test } from '../../../utils/text-formater';
+import { bbcode2html, html2bbcode } from '../../../utils/text-formater';
 import './text-editor.scss';
-// TODO: support [br]
+
 // TODO: 表情包
-// TODO: font size
+// TODO: 字号要调大一点
+// FIXME: 老站的code是inline element，新的code是block element。见test case #complex3 老站允许在code中字体加粗，调字号，甚至创建list。
+// TODO: 加链接的UX有点别扭，在没有选中东西的时候提醒用户选中东西。
+// TODO: santize html
+// TODO: 和谐词过滤
+// TODO: 圈人
+// TODO: 搞一个style2给私聊页面
+// TODO: 图片允许调大小
+
+// OTHER NOTES
+// FIXME: if run a test include <code>, then run a test without, the code style still apply (e.g.in default test suit, try test "code", then the next test case). It seems that there is something wrong in the state sync in quill package, fixing may need some time. However, since that we do not really need to update prop.content during runtime, the bug has low priority.
+
 // TODO: there are some lifecycle warnings with this component (e.g.omponentWillUpdate has been renamed), this warning is from the library Quill
-// FIXME: code test first, then other test will have code style
 // https://github.com/quilljs/quill/issues/2771
 // Thre ReactQuill is a react wrapper for Quill, it also has this warning: https://github.com/zenoamaro/react-quill/pull/531
 // however, the ReactQuill team is already working on fixing this, and the PR has already fixed the issue https://github.com/zenoamaro/react-quill/pull/549
 // As it seems that the maintainer plans to merge this PR soon, I would prefer to wait for a while first. If not, I will clone the reactQuill module and try fix it myself Q.Q
-
-// TODO: the editor supports upload local img, but probably we only want to accept link for web img.
-
-// TODO: TEST
-// [{ 'size': ['small', false, 'large', 'huge'] }], // fail -> use header for now
-// [{ 'color': [] }, { 'background': [] }],
-// ['bold', 'italic', 'underline', 'strike', 'blockquote'],
-// [{'list': 'ordered'}, {'list': 'bullet'}, {'indent': '-1'}, {'indent': '+1'}],
-// ['link', 'image'],
-// ['clean']
 
 export type textFormat = 'plaintext' | 'markdown' | 'bbcode';
 
@@ -69,7 +68,7 @@ export class TextEditor extends React.Component<{
     }
     const range = this.quillRef.getSelection();
     // TODO: use a common prompt element
-    const value = prompt('What is the image URL');
+    const value = prompt('输入图片URL');
     if (value) {
         this.quillRef.insertEmbed(range.index, 'image', value, 'user');
       }
@@ -115,13 +114,6 @@ export class TextEditor extends React.Component<{
       if (!isMarkdown) {
         const html = bbcode2html(content);
         return html;
-        // this.reactQuillRef.clip
-        // TODO
-        // if (this.sanitize) {
-          // this.content = this.domSanitizer.sanitize(SecurityContext.HTML, this.content);
-      // }
-      // const contents = this.quillEditor.clipboard.convert(this.content);
-      // this.quillEditor.setContents(contents, 'silent');
       }
     }
     return '';
